Create README.md
I want to write a comic, and I had a general idea of what I wanted the story to be, but then I realised I don't want to spend that much time on something unless it's truely meaningful. So now I've got a bunch of unplaced characters floating around in my head who have no story. -_-